ref(tables): promote the explore table shim into a shared DataTable
`views/explore/components/table.tsx` was a deprecated shim that reached
into `tables/gridEditable/styles` for its chrome. Nine more files had
started reaching in the same way, importing `GridStatus`, `GridRow` and
friends by name, so GridEditable's internals could not be restyled
without breaking explore, logs, profiling and issue details.

That chrome is now `components/tables/dataTable`, a `DataTable` compound
component owning the row tracks, scroll containment, head, row, cell and
status styling that sits between `@sentry/scraps/table` and its callers.
GridEditable composes it, the shim's callers use it directly, and the
shim is gone.

`gridEditable/styles.tsx` keeps only what is GridEditable's own: the
title/buttons header and the two prepend-column cells.

`ALLOWED_CELL_ACTIONS` had nothing to do with the table and moves to
`views/explore/components/cellActions`.

No visual change: every rule moved verbatim, and `align` on a head cell
is now handled once by `DataTable.HeadCell` instead of again in
multiQueryMode.

Continues DE-1392.
