+++ static/app/views/explore/multiQueryMode/queryVisualizations/table.tsx
@@ -1,6 +1,5 @@
 import {Fragment, useMemo} from 'react';
 import {useTheme} from '@emotion/react';
-import {css} from '@emotion/react';
 import styled from '@emotion/styled';
 
 import {Flex} from '@sentry/scraps/layout';
@@ -9,12 +8,7 @@ import {Tooltip} from '@sentry/scraps/tooltip';
 
 import {EmptyStateWarning} from 'sentry/components/emptyStateWarning';
 import {LoadingIndicator} from 'sentry/components/loadingIndicator';
-import type {ColumnAlign} from 'sentry/components/tables/gridEditable';
-import {
-  GridBodyCell,
-  GridHeadCell,
-  GridStatus,
-} from 'sentry/components/tables/gridEditable/styles';
+import {DataTable} from 'sentry/components/tables/dataTable';
 import {IconArrow} from 'sentry/icons/iconArrow';
 import {IconStack} from 'sentry/icons/iconStack';
 import {IconWarning} from 'sentry/icons/iconWarning';
@@ -27,7 +21,6 @@ import {
 } from 'sentry/utils/discover/fields';
 import {prettifyTagKey} from 'sentry/utils/fields';
 import {useLocation} from 'sentry/utils/useLocation';
-import {TableBody, TableHead, TableRow} from 'sentry/views/explore/components/table';
 import {Mode} from 'sentry/views/explore/contexts/pageParamsContext/mode';
 import type {AggregatesTableResult} from 'sentry/views/explore/hooks/useExploreAggregatesTable';
 import type {SpansTableResult} from 'sentry/views/explore/hooks/useExploreSpansTable';
@@ -109,8 +102,8 @@ function AggregatesTable({
         prefixColumnWidth="min-content"
         scrollable
       >
-        <TableHead>
-          <TableRow>
+        <DataTable.Head>
+          <DataTable.Row>
             <TableHeadCell isFirst={false}>
               <Flex align="center" gap="xs" />
             </TableHeadCell>
@@ -159,22 +152,22 @@ function AggregatesTable({
                 </TableHeadCell>
               );
             })}
-          </TableRow>
-        </TableHead>
-        <TableBody>
+          </DataTable.Row>
+        </DataTable.Head>
+        <DataTable.Body>
           {result.isPending ? (
-            <GridStatus>
+            <DataTable.Status>
               <LoadingIndicator />
-            </GridStatus>
+            </DataTable.Status>
           ) : result.isError ? (
-            <GridStatus>
+            <DataTable.Status>
               <IconWarning data-test-id="error-indicator" variant="muted" size="lg" />
-            </GridStatus>
+            </DataTable.Status>
           ) : result.isFetched && result.data?.length ? (
             result.data?.map((row, i) => {
               const target = getSamplesTargetAtIndex(index, [...queries], row, location);
               return (
-                <TableRow key={i}>
+                <DataTable.Row key={i}>
                   <TableBodyCell key={`samples-${i}`}>
                     {i < TOP_EVENTS_LIMIT && <TopResultsIndicator color={palette[i]!} />}
                     <Tooltip title={t('View Samples')} containerDisplayMode="flex">
@@ -196,17 +189,17 @@ function AggregatesTable({
                       </TableBodyCell>
                     );
                   })}
-                </TableRow>
+                </DataTable.Row>
               );
             })
           ) : (
-            <GridStatus>
+            <DataTable.Status>
               <EmptyStateWarning>
                 <p>{t('No spans found')}</p>
               </EmptyStateWarning>
-            </GridStatus>
+            </DataTable.Status>
           )}
-        </TableBody>
+        </DataTable.Body>
       </Table>
     </Fragment>
   );
@@ -240,8 +233,8 @@ function SpansTable({spansTableResult, query: queryParts, index}: SampleTablePro
         minimumColumnWidth={50}
         scrollable
       >
-        <TableHead>
-          <TableRow>
+        <DataTable.Head>
+          <DataTable.Row>
             {visibleFields.map((field, i) => {
               // Hide column names before alignment is determined
               if (result.isPending) {
@@ -278,20 +271,20 @@ function SpansTable({spansTableResult, query: queryParts, index}: SampleTablePro
                 </TableHeadCell>
               );
             })}
-          </TableRow>
-        </TableHead>
-        <TableBody>
+          </DataTable.Row>
+        </DataTable.Head>
+        <DataTable.Body>
           {result.isPending ? (
-            <GridStatus>
+            <DataTable.Status>
               <LoadingIndicator />
-            </GridStatus>
+            </DataTable.Status>
           ) : result.isError ? (
-            <GridStatus>
+            <DataTable.Status>
               <IconWarning data-test-id="error-indicator" variant="muted" size="lg" />
-            </GridStatus>
+            </DataTable.Status>
           ) : result.isFetched && result.data?.length ? (
             result.data?.map((row, i) => (
-              <TableRow key={i}>
+              <DataTable.Row key={i}>
                 {visibleFields.map((field, j) => {
                   return (
                     <TableBodyCell key={j}>
@@ -305,16 +298,16 @@ function SpansTable({spansTableResult, query: queryParts, index}: SampleTablePro
                     </TableBodyCell>
                   );
                 })}
-              </TableRow>
+              </DataTable.Row>
             ))
           ) : (
-            <GridStatus>
+            <DataTable.Status>
               <EmptyStateWarning>
                 <p>{t('No spans found')}</p>
               </EmptyStateWarning>
-            </GridStatus>
+            </DataTable.Status>
           )}
-        </TableBody>
+        </DataTable.Body>
       </Table>
     </Fragment>
   );
@@ -334,17 +327,12 @@ const StyledLink = styled(Link)`
   display: flex;
 `;
 
-const TableBodyCell = styled(GridBodyCell)`
+const TableBodyCell = styled(DataTable.Cell)`
   font-size: ${p => p.theme.font.size.sm};
   min-height: 12px;
 `;
 
-const TableHeadCell = styled(GridHeadCell)<{align?: ColumnAlign}>`
-  ${p =>
-    p.align &&
-    css`
-      justify-content: ${p.align};
-    `}
+const TableHeadCell = styled(DataTable.HeadCell)`
   font-size: ${p => p.theme.font.size.sm};
   height: 33px;
 `;
